+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Features
 - Update logic of selecting the default project and site to the most relevant one for the user on the Creating a JIRA issue page
+- Added filtering for options and error message in projects field on Create Jira issue page.
 
 ### Bug Fixes